+++ docs/src/conf.py
@@ -22,6 +22,7 @@ import sys
 import json
 import yaml
 from datetime import date
+import re
 
 
 # -- Project information -----------------------------------------------------
@@ -116,6 +117,13 @@ if not version:
     sys.exit(1)
 chart_version = version
 
+if re.match(r"^[0-9]+.[0-9]+.[0-9]+$", version):
+    # For releases, we download artifacts from GitHub Releases
+    download_url = f"https://github.com/digital-asset/decentralized-canton-sync/releases/download/v{version}"
+else:
+    # For snapshots, we download artifacts through the gcs proxy on the cluster
+    download_url = "/cn-release-bundles"
+
 # Sphinx does not allow something like ``|version|``
 # so instead we define a replacement that includes the formatting.
 rst_prolog = f"""
@@ -137,8 +145,8 @@ rst_prolog = f"""
 .. |chart_version_set| replace:: ``export CHART_VERSION={chart_version}``
 .. |image_tag_set| replace:: ``export IMAGE_TAG={version}``
 
-.. |bundle_download_link| replace:: :raw-html:`<a class="reference external" href="/cn-release-bundles/{version}_splice-node.tar.gz">Download Bundle</a>`
-.. |openapi_download_link| replace:: :raw-html:`<a class="reference external" href="/cn-release-bundles/{version}_openapi.tar.gz">Download OpenAPI specs</a>`
+.. |bundle_download_link| replace:: :raw-html:`<a class="reference external" href="{download_url}/{version}_splice-node.tar.gz">Download Bundle</a>`
+.. |openapi_download_link| replace:: :raw-html:`<a class="reference external" href="{download_url}/{version}_openapi.tar.gz">Download OpenAPI specs</a>`
 
 .. |canton_download_link| replace:: :raw-html:`<a class="reference external" href="https://digitalasset.jfrog.io/artifactory/canton-enterprise/canton-enterprise-{canton_version}.tar.gz">Download Canton enterprise</a>`
 """